feat: sourcemap api

+++ app/manage/model/dto/syslog.go
@@ -2,6 +2,6 @@ package dto
 
 // nsq 消息解析失败
 type ParseErrorParam struct {
-	Current  int `form:"current" binding:"number"`
-	PageSize int `form:"pageSize" binding:"number"`
+	Current  int64 `form:"current" binding:"number"`
+	PageSize int64 `form:"pageSize" binding:"number"`
 }
